Estimate Population Immunity, Reduction in R0, GMT, and Proportion of Non-Naive Individuals

This function estimates population immunity, the relative reduction in transmissibility (R0), the geometric mean titer (GMT), and the proportion of individuals with HI > 5 using bootstrap sampling and MCMC methods. It considers multiple age groups and various parameters related to immunity.

Usage

ImmuPop_timet_est(
  df,
  protect_c,
  protect_a,
  age_prop,
  contact_matrix,
  sim_num = 500
)

Arguments

df

Data frame containing individual data, including age group and raw titer values.

protect_c

Numeric vector of protection effect for children.

protect_a

Numeric vector of protection effect for adults.

age_prop

Numeric vector representing age group proportions in the population.

contact_matrix

Numeric matrix of contact frequencies between age groups.

sim_num

Number of bootstrap simulations (default = 500).

Value

A dataframe with baseline immunity estimates, including the median and 95

Examples

# Example usage of ImmuPop_timet_est
data("ImmuPop_raw_data")
data_example <- generate_data(ImmuPop_raw_data, cut_age = c(0, 18, 50, 100))

# Define age group proportions
age_prop <- c(0.2, 0.4, 0.4)

# Define contact matrix
contact_matrix <- matrix(c(22, 16, 15, 24, 28, 30, 18, 32, 35), nrow = 3, byrow = TRUE)

# Define protection effects for children and adults
protect_c <- c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.95)
protect_a <- c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.95)

# Data at a specific time point (e.g., time == 200)
data_example_t <- data_example[data_example$time == 200, ]

# Run the estimation function
result <- ImmuPop_timet_est(
  df = data_example_t,
  protect_c = protect_c,
  protect_a = protect_a,
  age_prop = age_prop,
  contact_matrix = contact_matrix,
  sim_num = 100
)

# Print the result
print(result)
#>   estimator     value    CI_lwr     CI_upr
#> 1 pop_immun 0.2140778 0.1768733  0.3905841
#> 2     RR_R0 0.2260930 0.2093160  0.3699402
#> 3       GMT 7.1000000 2.2950000 11.3050000
#> 4    prop_5 0.4000000 0.0000000  0.8000000