Generate Example Data from Raw Immunity Data

This function takes raw immunity data (including individual age and raw titer values), and generates a data frame with additional columns: age group and titer level. The age group is categorized based on the provided `cut_age` vector. The titer level is calculated using the formula `log2(raw_titer/5) + 1`.

Arguments

ImmuPop_raw_data

A data frame containing individual raw data, which must have: - `age`: individual age - `raw_titer`: raw titer values.

cut_age

A numeric vector indicating the breakpoints to categorize ages. For example, `cut_age = c(0, 18, 60, 100)` would create the following age groups: - [0, 18), [18, 60), [60, 100).

Value

A data frame containing the following columns: - `agegp1`: age group based on the cut points from `cut_age`. - `titer_level`: log2 transformed titer level calculated as `log2(raw_titer / 5) + 1`.

Examples

# Example usage of generate_data with a vector for cut_age
raw_data <- data.frame(
  age = c(5, 15, 25, 40, 60),
  raw_titer = c(10, 50, 80, 100, 200)
)
processed_data <- generate_data(raw_data, cut_age = c(0, 18, 60, 100))
print(processed_data)
#>   age raw_titer   agegp1 titer_level
#> 1   5        10   [0,18)    2.000000
#> 2  15        50   [0,18)    4.321928
#> 3  25        80  [18,60)    5.000000
#> 4  40       100  [18,60)    5.321928
#> 5  60       200 [60,100)    6.321928